+++ assets/src/pages/notificationPermissions.js
@@ -1,0 +1,5 @@
+import { initCore } from 'src/lib/init.js'
+import { initJsonEditor } from 'src/lib/json_editor_init.js'
+
+initCore()
+initJsonEditor()