+++ python/aristotle-bg-workers/aristotle_bg_workers/celery.py
@@ -1,6 +1,7 @@
 from __future__ import absolute_import, unicode_literals
 from celery import Celery
 
+
 # set the default Django settings module for the 'celery' program.
 # os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'aristotle_mdr.settings')
 
@@ -22,8 +23,10 @@ def debug_task(self):
     print('Request: {0!r}'.format(self.request))
 
 
+
 # from celery import signals
 # @signals.setup_logging.connect
 # def setup_celery_logging(**kwargs):
 #     pass
 # app.log.setup()
+